Use file system for scratch dir on BGQ for sp jobs
Because the size of the files in the scratch space can easily become too large to fit in the
ramdisk for the second part of the sp job which is the pcm solvation, in this case
create the scratch directory in the file system, utilize and manage  it.  Also, use the
latest version 5 of the qchem binary on vesta.